Share reservation activity helper
