+++ api/select-club.ts
@@ -5,6 +5,7 @@ import { getJwtPayload } from './verifyAuth.js';
 import { VercelRequest, VercelResponse } from '@vercel/node';
 import { DBUser, ReservationItem } from '../src/types.js';
 import sendEmail from './_sendEmail.js';
+import { isReservationActive } from '../src/utils/reservations.js';
 
 type ClubDocument = {
   name?: string;
@@ -22,17 +23,6 @@ const validationError = (message: string) => ({
     }
   ]
 });
-
-const isReservationActive = (reservation: ReservationItem, now = new Date()) => {
-  if (reservation.recurring) {
-    return true;
-  }
-
-  const reservationEndTime = new Date(reservation.date);
-  reservationEndTime.setHours(reservation.end_time, 0, 0, 0);
-
-  return reservationEndTime > now;
-};
 
 async function deleteActiveReservationsForUser(
   reservationsCollection: ReturnType<MongoClient['db']>['collection'],
