+++ api/users.ts
@@ -6,6 +6,7 @@ import { VercelRequest, VercelResponse } from '@vercel/node';
 import sendEmail from './_sendEmail.js';
 import { escapeHtml } from './_lib.js';
 import { ReservationItem } from '../src/types.js';
+import { isReservationActive } from '../src/utils/reservations.js';
 
 type ClubDocument = {
   name?: string;
@@ -14,17 +15,6 @@ type ClubDocument = {
 function getStatusLabel(status: string) {
   return status === 'active' ? 'aktiv' : 'inaktiv';
 }
-
-const isReservationActive = (reservation: ReservationItem, now = new Date()) => {
-  if (reservation.recurring) {
-    return true;
-  }
-
-  const reservationEndTime = new Date(reservation.date);
-  reservationEndTime.setHours(reservation.end_time, 0, 0, 0);
-
-  return reservationEndTime > now;
-};
 
 async function deleteActiveReservationsForUser(
   database: ReturnType<MongoClient['db']>,
